+++ src/main.rs
@@ -6,17 +6,14 @@ use clap::{App, Arg};
 
 mod gameloader;
 mod osspec;
-mod engine;
 mod jenkins;
 
 fn main() {
-    let result = jenkins::getBranchURL("master");
-    println!("{}", result.unwrap().url);
     let matches = App::new("Kakara Game Launcher").
         version("1.0-SNAPSHOT").author("Wyatt Jacob Herkamp <wherkamp@kingtux.me>").about("Launches the Kakara game").
-        arg(Arg::with_name("game").short("g").long("game").value_name("JAR_FILE").help("Takes the Kakara client").takes_value(true).required(true)).
+        arg(Arg::with_name("game").short("g").long("game").value_name("JAR_FILE").help("Takes the Kakara client").takes_value(true).required(false)).
         arg(Arg::with_name("engine").short("e").long("engine").value_name("JAR_FILE").help("Takes the Kakara Engine").takes_value(true).required(false)).
-        arg(Arg::with_name("dir").short("w").long("working_dir").value_name("WORKING_DIRECTORY").help("What is the working directory for Kakara").takes_value(true).required(true)).
+        arg(Arg::with_name("dir").short("w").long("working_dir").value_name("WORKING_DIRECTORY").help("What is the working directory for Kakara").takes_value(true).required(false)).
         get_matches();
     let game_jar = matches.value_of("game").unwrap_or("client.jar");
     let engine_jar: String;
@@ -25,8 +22,11 @@ fn main() {
     if matches.is_present("engine") {
         engine_jar = matches.value_of("engine").unwrap_or("engine.jar").parse().unwrap();
     } else {
+        let result = jenkins::getBranchURL("master");
+        let job = result.unwrap();
+        let engine1 = jenkins::get_build_url(job);
         let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-        let s = runtime.block_on(engine::download_latest_build(working_directory));
+        let s = runtime.block_on(jenkins::download_engine_jar(&*engine1));
         if s == "" {
             // If the engine version was not found.
             println!("[ERROR] Unable to download engine version. Please provide an engine build with --engine");
